Moved all UI components under ui/ directory.
