+++ docs/coverage.out
@@ -1164,15 +1164,15 @@ github.com/gomlx/gomlx/ml/layers/regularizers/regularizers.go:154:		ConstantL1		
 github.com/gomlx/gomlx/ml/train/batchnorm.go:25:				BatchNormalizationAveragesUpdate		0.0%
 github.com/gomlx/gomlx/ml/train/batchnorm.go:59:				batchNormAveragesStep				0.0%
 github.com/gomlx/gomlx/ml/train/batchnorm.go:70:				batchNormsAverageStepGraphFn			0.0%
-github.com/gomlx/gomlx/ml/train/commandline/commandline.go:27:			ReportEval					100.0%
-github.com/gomlx/gomlx/ml/train/commandline/contextsettings.go:42:		ParseContextSettings				68.1%
-github.com/gomlx/gomlx/ml/train/commandline/contextsettings.go:165:		CreateContextSettingsFlag			91.7%
-github.com/gomlx/gomlx/ml/train/commandline/contextsettings.go:189:		SprintContextSettings				85.7%
-github.com/gomlx/gomlx/ml/train/commandline/progressbar.go:40:			Write						100.0%
-github.com/gomlx/gomlx/ml/train/commandline/progressbar.go:49:			onStart						87.5%
-github.com/gomlx/gomlx/ml/train/commandline/progressbar.go:70:			onStep						63.6%
-github.com/gomlx/gomlx/ml/train/commandline/progressbar.go:114:			onEnd						100.0%
-github.com/gomlx/gomlx/ml/train/commandline/progressbar.go:142:			AttachProgressBar				100.0%
+github.com/gomlx/gomlx/ui/commandline/commandline.go:27:			ReportEval					100.0%
+github.com/gomlx/gomlx/ui/commandline/contextsettings.go:42:		ParseContextSettings				68.1%
+github.com/gomlx/gomlx/ui/commandline/contextsettings.go:165:		CreateContextSettingsFlag			91.7%
+github.com/gomlx/gomlx/ui/commandline/contextsettings.go:189:		SprintContextSettings				85.7%
+github.com/gomlx/gomlx/ui/commandline/progressbar.go:40:			Write						100.0%
+github.com/gomlx/gomlx/ui/commandline/progressbar.go:49:			onStart						87.5%
+github.com/gomlx/gomlx/ui/commandline/progressbar.go:70:			onStep						63.6%
+github.com/gomlx/gomlx/ui/commandline/progressbar.go:114:			onEnd						100.0%
+github.com/gomlx/gomlx/ui/commandline/progressbar.go:142:			AttachProgressBar				100.0%
 github.com/gomlx/gomlx/ml/train/loop.go:94:					NewLoop						100.0%
 github.com/gomlx/gomlx/ml/train/loop.go:117:					GetTrainLastStepVar				100.0%
 github.com/gomlx/gomlx/ml/train/loop.go:127:					start						71.4%
