+++ docs/notebooks/cifar.html
@@ -7804,7 +7804,7 @@ drwxr-x--- 2 janpf janpf 4.0K Aug  1 14:24 fnn_layer
     <span class="s2">"github.com/gomlx/gomlx/ml/layers/fnn"</span>
     <span class="s2">"github.com/gomlx/gomlx/ml/layers/kan"</span>
     <span class="s2">"github.com/gomlx/gomlx/ml/layers/regularizers"</span>
-    <span class="s2">"github.com/gomlx/gomlx/ml/train/commandline"</span>
+    <span class="s2">"github.com/gomlx/gomlx/ui/commandline"</span>
     <span class="s2">"github.com/gomlx/gomlx/ml/train/optimizers"</span>
     <span class="s2">"github.com/gomlx/gomlx/examples/cifar"</span>
     <span class="s2">"github.com/gomlx/gomlx/ml/context"</span>
