+++ docs/notebooks/fft.html
@@ -7564,7 +7564,7 @@ The goal is to do an <code>InverseRealFFT</code> by gradient descent.</p>
     <span class="o">.</span> <span class="s2">"github.com/gomlx/exceptions"</span>
     <span class="s2">"github.com/gomlx/gomlx/types/tensors"</span>
     <span class="s2">"github.com/gomlx/gomlx/types/xslices"</span>
-    <span class="n">mg</span> <span class="s2">"github.com/gomlx/gomlx/examples/notebook/gonb/margaid"</span>
+    <span class="n">mg</span> <span class="s2">"github.com/gomlx/gomlx/ui/gonb/margaid"</span>
     <span class="s2">"github.com/janpfeifer/gonb/gonbui"</span>
 
     <span class="n">_</span> <span class="s2">"github.com/gomlx/gomlx/backends/xla"</span>
