+++ docs/notebooks/uci-adult.html
@@ -7706,7 +7706,7 @@ drwxr-x--- 2 janpf janpf 4.0K Aug  9 11:51 kan_model
     <span class="s2">"github.com/gomlx/gomlx/ml/layers/fnn"</span>
     <span class="s2">"github.com/gomlx/gomlx/ml/layers/kan"</span>
     <span class="s2">"github.com/gomlx/gomlx/ml/layers/regularizers"</span>
-    <span class="s2">"github.com/gomlx/gomlx/ml/train/commandline"</span>
+    <span class="s2">"github.com/gomlx/gomlx/ui/commandline"</span>
     <span class="s2">"github.com/gomlx/gomlx/ml/train/optimizers"</span>
 	<span class="s2">"github.com/janpfeifer/must"</span>    
 <span class="p">)</span>
@@ -8104,7 +8104,7 @@ it is working.</p>
 <div class="highlight hl-ipython3"><pre><span></span><span class="kn">import</span> <span class="p">(</span>
     <span class="s2">"fmt"</span>
     <span class="s2">"time"</span>
-    <span class="s2">"github.com/gomlx/gomlx/examples/notebook/gonb/plotly"</span>
+    <span class="s2">"github.com/gomlx/gomlx/ui/gonb/plotly"</span>
 <span class="p">)</span>
 
 <span class="n">var</span> <span class="p">(</span>
