+++ examples/ogbnmag/demo/main.go
@@ -12,9 +12,9 @@ import (
 	"github.com/gomlx/gomlx/ml/layers/activations"
 	"github.com/gomlx/gomlx/ml/layers/kan"
 	"github.com/gomlx/gomlx/ml/layers/regularizers"
-	"github.com/gomlx/gomlx/ml/train/commandline"
 	"github.com/gomlx/gomlx/ml/train/optimizers"
 	"github.com/gomlx/gomlx/ml/train/optimizers/cosineschedule"
+	"github.com/gomlx/gomlx/ui/commandline"
 	"github.com/gomlx/gomlx/ui/fyneui"
 	"github.com/janpfeifer/must"
 	"k8s.io/klog/v2"
