+++ examples/oxfordflowers102/diffusion/train/main.go
@@ -4,7 +4,7 @@ import (
 	"flag"
 	"github.com/gomlx/exceptions"
 	"github.com/gomlx/gomlx/examples/oxfordflowers102/diffusion"
-	"github.com/gomlx/gomlx/ml/train/commandline"
+	"github.com/gomlx/gomlx/ui/commandline"
 	"github.com/janpfeifer/must"
 	"k8s.io/klog/v2"
