+++ ui/bashkernel/bashkernel.go
