+++ ui/gonb/README.md
@@ -1,0 +1,5 @@
+# [GoNB](https://github.com/janpfeifer/gonb) Tools
+
+[GoNB](https://github.com/janpfeifer/gonb) is a Jupyter Notebook kernel for Go, and the packages in this directory
+allows for plotting using different libraries that will finally output to HTML+Javascript in the output cell of a 
+notebook.
