+++ ui/gonb/margaid/margaid.go
@@ -42,10 +42,10 @@ import (
 	"encoding/json"
 	"fmt"
 	mg "github.com/erkkah/margaid"
-	stdplots "github.com/gomlx/gomlx/examples/notebook/gonb/plots"
 	"github.com/gomlx/gomlx/ml/train"
 	"github.com/gomlx/gomlx/types/tensors"
 	"github.com/gomlx/gomlx/types/xslices"
+	stdplots "github.com/gomlx/gomlx/ui/plots"
 	"github.com/janpfeifer/gonb/gonbui"
 	"github.com/pkg/errors"
 	"io"
