+++ ui/gonb/plotly/plotly.go
@@ -15,13 +15,13 @@ package plotly
 import (
 	"fmt"
 	grob "github.com/MetalBlueberry/go-plotly/graph_objects"
-	"github.com/gomlx/gomlx/examples/notebook/gonb/plots"
 	"github.com/gomlx/gomlx/ml/context/checkpoints"
 	"github.com/gomlx/gomlx/ml/data"
 	"github.com/gomlx/gomlx/ml/train"
 	"github.com/gomlx/gomlx/types"
 	"github.com/gomlx/gomlx/types/tensors"
 	"github.com/gomlx/gomlx/types/xslices"
+	"github.com/gomlx/gomlx/ui/plots"
 	"github.com/janpfeifer/gonb/gonbui"
 	"github.com/janpfeifer/gonb/gonbui/dom"
 	gonbplotly "github.com/janpfeifer/gonb/gonbui/plotly"
